MAJOR BUGS DISCOVERED AND ONE FIXED
I found tons bugs in rootShell maybe, this bug is triggered when you
type [some random char][command][or some random char here] which is:
xdir extension instead of dir extension. This method and method
cd_to_dir are vulnerable to this type of command execution. cd_to_dir
method is now free of this bug but dir_extension method is yet to be
free of this bug and it is anticipated that few other methods are weak
to this bug too. Most are in effect but some of them were still stable
enough even though they were not showing good responses.